+++ CodexLib/Components/AbilityTargetHasFactExcept.cs
@@ -4,6 +4,7 @@ using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.Blueprints.Root;
 using Kingmaker.EntitySystem.Entities;
 using Kingmaker.UI.Log;
+using Kingmaker.UI.Models.Log;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components.Base;